🧹 separation of network interface from the type business
