+++ src/EcologicalNetworks.jl
@@ -55,8 +55,9 @@ include(joinpath(".", "interfaces/abstractarray.jl"))
 include(joinpath(".", "interfaces/iteration.jl"))
 
 # General useful manipulations
-include(joinpath(".", "types/utilities.jl"))
-include(joinpath(".", "types/comparisons.jl"))
+include(joinpath(".", "utilities/comparisons.jl"))
+include(joinpath(".", "utilities/overloads.jl"))
+include(joinpath(".", "utilities/utilities.jl"))
 export species, interactions, has_interaction, richness, nodiagonal, nodiagonal!
 
 # Degree
